requirements.txt
cups==0.0.6
Flask==3.0.2
Flask-Cors==4.0.0
pyusb==1.2.1
